requirements.txt
tensorflow
tensorflow-addons
numpy
scikit-image
matplotlib
imgaug
Pillow
pynrrd
scipy
tqdm